+++ app/[lang]/(app)/(home)/(tablet)/restaurant/salons/components/merge-table/MergeTableModal.tsx
@@ -52,7 +52,13 @@ export default function MergeTableModal({
  });
 
  return (
-  <Dialog open={open} onOpenChange={changeOpen}>
+  <Dialog
+   open={open}
+   onOpenChange={(newValue) => {
+    if (isPending) return;
+    changeOpen(newValue);
+   }}
+  >
    <DialogContent className='gap-0 p-0 sm:max-w-3xl'>
     <DialogHeader className='p-4 py-6'>
      <DialogTitle>{dic.mergeTableModal.title}</DialogTitle>