+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 # 2.1.1
 - update new order cash bank type when pay type is check (hra-done)
+- show two tables(normal-mode) in mobile resolution (hra-done)
 
 # 2.1.0
 ### features